Moved child process termination signal handler to a new module.
